feat: Add native esp-idf template
Tested with generating a Vue3 app for esp32, using esp-idf v5.3.2
This is perfectly functional, but there's always room for improvement.
Published in the hope it may be useful for others.

Note: no examples for esp-idf. Not sure where best to put them.

mini example
```
 #include <esp_http_server.h>
 #include "your_generated_file.h"
void start_http_server(void)
{
	httpd_handle_t httpd;
	httpd_config_t config = HTTPD_DEFAULT_CONFIG();
	config.max_uri_handlers = SVELTEESP32_COUNT + xxx;

	LOGI("Starting server on port: '%d'", config.server_port);
	ESP_ERROR_CHECK(httpd_start(&httpd, &config));
	initSvelteStaticFiles(httpd);
	// any other handlers you need....
}
```

Tested with and without etag, and with and without gzip.

Signed-off-by: Karl Palsson <karl.palsson@marel.com>